rework and add rule name/order validators
